Build on every push and pull request via GitHub Actions (#235)
* refactor: Update package imports for internal utils

This commit updates the package imports for internal utility files to reflect the changes in the file structure. The imports have been changed from "github.com/WangYihang/Platypus/internal/util" to "github.com/WangYihang/Platypus/internal/utils". This change ensures that the codebase is aligned with the new file structure and improves code organization.

Refactor the package imports for internal utility files to reflect the changes in the file structure

* refactor: Remove unused CodeQL analysis workflow and release workflow

* docs: fix invalid image path (#234)

* refactor: Remove unused go-bindata-assetfs package import

The go.mod file was modified to remove the import statement for the go-bindata-assetfs package, as it is no longer needed in the codebase.